pyproject.toml
[project]
name = "yapytr"
version = "0.1.0"
authors = [
    {name = "ExploracuriousAlex"},
]
description = "Use TradeRepublic in terminal"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["Trade Republic"]
license = {text = "MIT License"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Financial :: Investment"
]
dependencies = [
    "coloredlogs",
    "ecdsa",
    "packaging",
    "pathvalidate",
    "pygments",
    "requests_futures",
    "shtab",
    "websockets>=10.1"
]

[project.scripts]
yapytr = "yapytr.__main__:main"